requirements.txt
flask==1.1.2
werkzeug==1.0.1
tensorflow==2.4.1
keras==2.4.3
numpy==1.19.5
Pillow==8.1.0
google-cloud-firestore==2.0.2
scikit-learn==0.24.1